+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -9,9 +9,9 @@ import java.time.LocalDate;
  */
 @Data
 public class Film {
-    Long id;
-    String name;
-    String description;
-    LocalDate releaseDate;
-    int duration;
+    private Long id;
+    private String name;
+    private String description;
+    private LocalDate releaseDate;
+    private int duration;
 }
