+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -6,9 +6,9 @@ import java.time.LocalDate;
 
 @Data
 public class User {
-    Long id;
-    String email;
-    String login;
-    String name;
-    LocalDate birthday;
+    private Long id;
+    private String email;
+    private String login;
+    private String name;
+    private LocalDate birthday;
 }
