+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -1,0 +1,15 @@
+package ru.yandex.practicum.filmorate.storage;
+
+import org.springframework.web.bind.annotation.RequestBody;
+import ru.yandex.practicum.filmorate.model.Film;
+
+import java.util.Collection;
+
+public interface FilmStorage {
+
+    public Collection<Film> getAll();
+
+    public Film create(@RequestBody Film film);
+
+    public Film update(@RequestBody Film film);
+}
