+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -1,0 +1,14 @@
+package ru.yandex.practicum.filmorate.storage;
+
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.util.Collection;
+
+public interface UserStorage {
+
+    public Collection<User> getAll();
+
+    public User create(User user);
+
+    public User update(User user);
+}
